requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-silk==4.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
